Battery
Added a battery percept.

Removed useless methods such as getBatteryCapacity in robot because it
was not needed.


Former-commit-id: 9793808cdb76ad0d5d574c81a2814df0b7804a16 [formerly f34fde41b1f1393868d6eb8c8582280c624b8f5b]
Former-commit-id: 1efa3dfa77f77acf79ba3390ba55d8c25f28220e

+++ bw4t/bw4t-server/src/main/java/nl/tudelft/bw4t/robots/Battery.java
@@ -52,12 +52,11 @@ public class Battery {
     }
     
     /**
-     * Returns percentage. TODO: make this function more accurate.
-     * Cast to double and reformat maybe?
+     * Returns percentage.
      * @return the percentage of battery power left.
      */
-    public int getPercentage() {
-        return (int) ((this.current * 100.0) / ((double) this.max));
+    public double getPercentage() {
+        return (this.current * 100.0) / this.max;
     }
     
     /**
